requirements.txt
kkccxt>=4.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
rich>=13.0.0
requests>=2.28.0
httpx>=0.24.0
ta>=0.11.0
